Win fixes

+++ src/lib/ndpi_community_id.c
@@ -70,7 +70,7 @@
 
 /* **************************************************** */
 
-static ssize_t ndpi_community_id_buf_copy(u_int8_t * const dst, const void * const src, ssize_t len) {
+static u_int16_t ndpi_community_id_buf_copy(u_int8_t * const dst, const void * const src, u_int16_t len) {
   if(src)
     memcpy(dst, src, len);
   else
